+++ packages/datafile-manager/src/reactNativeAsyncStorageCache.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import AsyncStorage from '@react-native-community/async-storage';
+import AsyncStorage from '@react-native-async-storage/async-storage';
 
 import PersistentKeyValueCache from './persistentKeyValueCache';
